Extract useSubscribeScheduleAutoMode() from useSubscribeScheduleAutoModeState()
