+++ src/api/index.ts
@@ -3,3 +3,4 @@ export * from "./use-state-subscription";
 export * from "./use-trace_ids-subscription";
 export * from "./use-continuous-enabled-subscription";
 export * from "./use-schedule-auto-mode-state-subscription";
+export * from "./use-schedule-auto-mode-subscription";
