+++ src/components/auto-mode-button/AutoModeButton.vue
@@ -10,6 +10,6 @@
 import ButtonError from "./modes/error/Button.vue";
 import ButtonOn from "./modes/on/Button.vue";
 import ButtonOff from "./modes/off/Button.vue";
-import { useSubscribeScheduleAutoModeState } from "@/api";
-const { autoMode } = await useSubscribeScheduleAutoModeState();
+import { useSubscribeScheduleAutoMode } from "@/api";
+const { autoMode } = await useSubscribeScheduleAutoMode();
 </script>
